Add palindrome example and integrate into project
Added `Palidrom.cpp` to demonstrate palindrome checking logic.
Updated `Basiskurs.vcxproj` and `Basiskurs.vcxproj.filters` to
include the new file. Modified `Main.cpp` to activate the example
when `AKTIVE_BEISPIEL` is set to 6. Implemented a loop-based
program in `Palidrom.cpp` that checks if a user-input word is a
palindrome using the `istPalidrom` function.