Add parameterized constructor and fix example activation
Implemented a parameterized constructor in `Parameterkonstruktor.cpp`
for the `Car` class, along with a default constructor. Both constructors
print messages to indicate when they are called. Updated the `main`
function to demonstrate the usage of both constructors.

Corrected a typo in `Main.cpp` by replacing `KJopierkontruktor.cpp`
with `Kopierkonstruktor.cpp`. Changed `#define AKTIVE_BEISPIEL` to `3`
to activate the `Parameterkonstruktor.cpp` example.