Update output formatting and card color in main()
Modified the `use()` method in the `NumberCards` class to include a newline character at the beginning of the output for better formatting.

Changed the default color of `NumberCards` in the `main()` function from `BLAU` to `GELB`.

Added a call to the `use()` method in the `main()` function to display card details for each card in the deck.