Add Node counter to track created Node instances
Introduced a static `Node::counter` variable to the `Node` class
to track the total number of `Node` objects created. Incremented
the counter in the `getNewnode` and `insertFront` functions
whenever a new `Node` is instantiated.

Updated the `main` function to print the value of `Node::counter`
to provide visibility into the total number of nodes created
during program execution.