Remove `main` function from `QueueMitKlasse.cpp`
The `main` function was removed from the `QueueMitKlasse.cpp` file.
It previously demonstrated basic usage of the `Stack` class,
including enqueuing the value `10` and displaying the stack's
contents. This change indicates that the file is no longer
intended to serve as the program's entry point, or the `main`
function has been relocated to another file.