Replace Palidrom.cpp with QueueStackPalidrom.cpp
Palidrom.cpp was removed and replaced with QueueStackPalidrom.cpp.
The new file contains identical functionality, with no changes
to the logic. Updates were made to `Basiskurs.vcxproj` and
`Basiskurs.vcxproj.filters` to reflect this replacement.

In `Main.cpp`, the inclusion of Palidrom.cpp was replaced with
QueueStackPalidrom.cpp for `AKTIVE_BEISPIEL == 10`. The new file
was categorized under `Source Files\Woche06\Beispiele` in the
project filters.