Add VectorAufgabe01 example and fix iterator bug
Added a new example `VectorAufgabe01.cpp` demonstrating the use of 2D vectors. Updated `Basiskurs.vcxproj` and `Basiskurs.vcxproj.filters` to include the new file. Modified `Main.cpp` to activate the new example when `AKTIVE_BEISPIEL` is set to 22.

Fixed a bug in `VectorIterator.cpp` where the wrong iterator variable was used in a `for` loop. Note: `VectorAufgabe01.cpp` contains a bug in the `for` loop that needs to be addressed.